refactor: 파일 구조 정리
- 파일 구조 정리 : 패키지 이동 및 파일 합치기
- 어노테이션 추가 : Getter, Setter, RequiredArgsConstructor
