+++ BEF/src/main/java/com/example/BEF/Course/Service/CourseService.java
@@ -6,32 +6,29 @@ import com.example.BEF.Course.DTO.CourseLocationRes;
 import com.example.BEF.Course.DTO.CourseSaveRes;
 import com.example.BEF.Course.Domain.Course;
 import com.example.BEF.Course.Domain.UserCourse;
+import com.example.BEF.Course.Repository.CourseRepository;
+import com.example.BEF.Course.Repository.UserCourseRepository;
 import com.example.BEF.Disabled.Domain.Disabled;
-import com.example.BEF.Disabled.Service.DisabledRepository;
+import com.example.BEF.Disabled.Repository.DisabledRepository;
 import com.example.BEF.Location.DTO.LocationInfoRes;
 import com.example.BEF.Location.DTO.UserLocationRes;
 import com.example.BEF.Location.Domain.Location;
-import com.example.BEF.Location.Service.LocationRepository;
+import com.example.BEF.Location.Repository.LocationRepository;
 import com.example.BEF.User.Domain.User;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class CourseService {
-    @Autowired
-    CourseRepository courseRepository;
 
-    @Autowired
-    UserCourseRepository userCourseRepository;
-
-    @Autowired
-    LocationRepository locationRepository;
-
-    @Autowired
-    DisabledRepository disabledRepository;
+    private final CourseRepository courseRepository;
+    private final UserCourseRepository userCourseRepository;
+    private final LocationRepository locationRepository;
+    private final DisabledRepository disabledRepository;
 
     // 코스 생성
     public CourseInfoRes createUserCourse(User user, String name, String description) {
