+++ BEF/src/main/java/com/example/BEF/Location/Repository/LocationRepository.java
@@ -1,4 +1,4 @@
-package com.example.BEF.VoiceSearch.Repository;
+package com.example.BEF.Location.Repository;
 
 import com.example.BEF.Location.Domain.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,7 +7,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface DescriptionRepository extends JpaRepository<Location, Long> {
+public interface LocationRepository extends JpaRepository<Location, Long> {
+    Location findLocationByContentId(Long contentId);
+
+    List<Location> findByAddrContainingAndAddrContaining(String state, String city);
+
     // 개요 컬럼에서 키워드가 존재하면 해당 개요 컬럼이 있는 튜플 전부 반환
     @Query("SELECT l FROM Location l WHERE l.description LIKE %:keyword%")
     List<Location>findDescription(@Param("keyword") String keyword);
