+++ BEF/src/main/java/com/example/BEF/Search/Client/OpenAIClient.java
@@ -1,9 +1,9 @@
-package com.example.BEF.VoiceSearch.Client;
+package com.example.BEF.Search.Client;
 
 
-import com.example.BEF.VoiceSearch.Config.OpenAIClientConfig;
-import com.example.BEF.VoiceSearch.DTO.WhisperTranscriptionRequest;
-import com.example.BEF.VoiceSearch.DTO.WhisperTranscriptionResponse;
+import com.example.BEF.Search.Config.OpenAIClientConfig;
+import com.example.BEF.Search.DTO.WhisperTranscriptionRequest;
+import com.example.BEF.Search.DTO.WhisperTranscriptionResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
