+++ BEF/src/main/java/com/example/BEF/Search/DTO/WhisperTranscriptionRequest.java
@@ -1,4 +1,4 @@
-package com.example.BEF.VoiceSearch.DTO;
+package com.example.BEF.Search.DTO;
 
 import lombok.Builder;
 import lombok.Data;
